[front] Refactor and relocate server-specific utilities to `tools` directory
